made changes to email sender
